Shift down binary outputs of segmentgestalt.m by 1
The plotted figure, which seems to have correct positions for clangs and segments, does not correspond with the function outputs. The binary output of the clangs and segments needs to be circularly shifted down by 1 to match the visualisation of the segments.

This fix makes the binary outputs and visualisations match and also increase the precision, recall, and F-measure of the algorithm (I have tried it with the Essen collection).